backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings>=2.3.0
python-dotenv
orjson
openai>=1.30.0
langchain>=0.2.6
langchain-community>=0.2.6
langchain-openai>=0.1.8
faiss-cpu
numpy
rank-bm25
PyMuPDF
pillow
pytesseract
python-multipart
redis
tiktoken